Refactoring: move conflict prevention lock codes (#351)
* move code

* add

* add

* rename directory

* rename test directory

* add

* rename test

* add

* fix wrapper test

* fix ruff
